test(countdown): mocking out the Sleep() call for a speedup
Since we still need to make sure we're sleeping, we're going to validate
the number of calls being made. That said, this is normally where I
would stop, if I even made it this far (in terms of actually making sure
sleep was called 3 times). Maybe I'm not being careful enough in my
testing, but I might also just be getting caught up on the fact that a
toy doesn't map to the real world 1:1.